+++ src/core/client/utility/texture.ts
@@ -32,6 +32,34 @@ export function loadTexture(dictionary: string) {
     });
 }
 
+export function drawTexture2D(
+    dictionary: string,
+    name: string,
+    position: alt.IVector2,
+    scale: number = 1,
+    opacity: number = 255
+) {
+    const identifier = `${dictionary}${name}`;
+    if (!textureData[identifier]) {
+        const [_, width, height] = native.getActiveScreenResolution(0, 0);
+        const resolution = native.getTextureResolution(dictionary, name);
+        textureData[identifier] = {
+            x: resolution.x / width,
+            y: resolution.y / height
+        };
+    }
+
+    const texture = textureData[identifier];
+    if (!texture) {
+        return;
+    }
+
+    const width = texture.x * scale;
+    const height = texture.y * scale;
+
+    native.drawSprite(dictionary, name, position.x, position.y, width, height, 0, 255, 255, 255, opacity, false);
+}
+
 export function drawTexture(dictionary: string, name: string, position: alt.Vector3, scale: number = 1) {
     const identifier = `${dictionary}${name}`;
     if (!textureData[identifier]) {