+++ documentation/changelog.md
@@ -58,6 +58,7 @@ description: List of all the changes and current versions.
 + Fixed Entering / Exiting Vehicle while Typing
 + Added Lock State for HUD
 + Added Time to HUD
++ Added Wanted Level(s)
 ```
 
 ## 1.8.2 - April 19 - May 14, 2021